+++ R/sps.R
@@ -1,68 +1,5 @@
-#---- Inclusion probability ----
-# Internal helpers
-pi <- function(x, n) {
-  x / sum(x) * n
-}
-
-.inclusion_prob <- function(x, n) {
-  res <- pi(x, n)
-  repeat {
-    to_ta <- which(res > 1)
-    if (length(to_ta) == 0L) break
-    res[to_ta] <- 1
-    keep_ts <- which(res < 1)
-    res[keep_ts] <- pi(x[keep_ts], n - length(x) + length(keep_ts))
-  }
-  res
-}
-
-.inclusion_prob_list <- function(x, n, s) {
-  if (not_strict_positive_vector(x)) {
-    stop(
-      gettext("'x' must be a strictly positive and finite numeric vector")
-    )
-  }
-  n <- trunc(n)
-  if (not_positive_vector(n)) {
-    stop(
-      gettext("'n' must be a positive and finite numeric vector")
-    )
-  }
-  s <- as.factor(s)
-  if (length(x) != length(s)) {
-    stop(
-      gettext("'x' and 's' must be the same length")
-    )
-  }
-  if (length(n) != nlevels(s)) {
-    stop(
-      gettext("'n' must have a single sample size for each level in 's'")
-    )
-  }
-  if (anyNA(s)) {
-    stop(
-      gettext("'s' cannot contain NAs")
-    )
-  }
-  # bins isn't wide enough by default to catch factors with unused levels
-  # at the end
-  if (any(tabulate(s, nbins = nlevels(s)) < n)) {
-    stop(
-      gettext("sample size 'n' is greater than or equal to population size")
-    )
-  }
-  Map(.inclusion_prob, split(x, s), n)
-}
-
-# Exported 
-inclusion_prob <- function(x, n, s = gl(1, length(x))) {
-  res <- unsplit(.inclusion_prob_list(x, n, s), s)
-  attributes(res) <- NULL # unsplit() mangles attributes
-  res
-}
-
-#---- Sampling ----
-# Internal helpers
+#---- Internal helpers ----
+# Sequential Poisson sampling
 .sps <- function(p, n, prn = NULL) {
   ts <- p < 1
   ta <- which(!ts)
@@ -95,7 +32,7 @@ inclusion_prob <- function(x, n, s = gl(1, length(x))) {
   )
 }
 
-# Exported
+#---- Stratified sampling ----
 stratify <- function(f) {
   f <- match.fun(f)
   # return function
@@ -113,14 +50,12 @@ stratify <- function(f) {
           gettext("'prn' must be a numeric vector between 0 and 1")
         )
       }
-    }
-    prn <- if (!is.null(prn)) {
-      split(prn, s)
+      prn <- split(prn, s)
     } else {
-      vector("list", nlevels(s))
+      prn <- vector("list", nlevels(s))
     }
     samp <- Map(f, .inclusion_prob_list(x, n, s), n, prn)
-    res <- unlist(Map(`[`, split(seq_along(x), s), samp), use.name = FALSE)
+    res <- unlist(Map(`[`, split(seq_along(x), s), samp), use.names = FALSE)
     weights <- unlist(lapply(samp, weights), use.names = FALSE)
     levels <- unlist(lapply(samp, levels), use.names = FALSE)
     ord <- order(res)
@@ -136,151 +71,6 @@ stratify <- function(f) {
 sps <- stratify(.sps)
 
 ps <- stratify(.ps)
-
-#---- Allocation ----
-# Internal helpers
-# Rounding methods
-largest_remainder <- function(p, n, initial) {
-  p <- p / sum(p)
-  np <- n * p
-  npf <- floor(np)
-  remainder <- rank(npf - np, ties.method = "first") <= n - sum(npf)
-  initial + npf + remainder
-}
-
-highest_averages <- function(divisor) {
-  f <- switch(
-    divisor,
-    "D'Hondt" = function(x) x + 1,
-    "Webster" = function(x) x + 0.5,
-    "Imperiali" = function(x) x + 2,
-    "Huntington-Hill" = function(x) sqrt(x * (x + 1)),
-    "Danish" = function(x) x + 1 / 3,
-    "Adams" = function(x) x,
-    "Dean" = function(x) x * (x + 1) / (x + 0.5)
-  )
-  # return function
-  res <- function(p, n, initial) {
-    if (is.null(names(initial))) {
-      names(initial) <- names(p)
-    }
-    while (n > 0) {
-      i <- which.max(p / f(initial))
-      initial[i] <- initial[i] + 1
-      n <- n - 1
-    }
-    initial
-  }
-  environment(res) <- list2env(list(f = f), parent = getNamespace("sps"))
-  res
-}
-
-# Exported
-# Expected coverage
-expected_coverage <- function(x, N, s = gl(1, length(x))) {
-  if (not_strict_positive_vector(x)) {
-    stop(
-      gettext("'x' must be a strictly positive and finite numeric vector")
-    )
-  }
-  N <- trunc(N)
-  if (not_positive_number(N)) {
-    stop(
-      gettext("'N' must be a positive and finite number")
-    )
-  }
-  if (N > length(x)) {
-    stop(
-      gettext("sample size 'N' is greater than or equal to population size")
-    )
-  }
-  s <- as.factor(s)
-  if (length(x) != length(s)) {
-    stop(
-      gettext("'x' and 's' must be the same length")
-    )
-  }
-  if (anyNA(s)) {
-    stop(
-      gettext("'s' cannot contain NAs")
-    )
-  }
-  p <- split(log(1 - .inclusion_prob(x, N)), s)
-  sum(1 - vapply(p, function(x) exp(sum(x)), numeric(1L)))
-}
-
-# Proportional allocation
-prop_allocation <- function(
-    x, N, s = gl(1, length(x)), min = 0, 
-    method = c("Largest-remainder", "D'Hondt", "Webster", "Imperiali", 
-               "Huntington-Hill", "Danish", "Adams", "Dean")  
-) {
-  if (not_strict_positive_vector(x)) {
-    stop(
-      gettext("'x' must be a strictly positive and finite numeric vector")
-    )
-  }
-  N <- trunc(N)
-  if (not_positive_number(N)) {
-    stop(
-      gettext("'N' must be a positive and finite number")
-    )
-  }
-  min <- trunc(min)
-  if (not_positive_number(min)) {
-    stop(
-      gettext("'min' must be a positive and finite number")
-    )
-  }
-  if (N > length(x)) {
-    stop(
-      gettext("sample size 'N' is greater than or equal to population size")
-    )
-  }
-  s <- as.factor(s)
-  if (length(x) != length(s)) {
-    stop(
-      gettext("'x' and 's' must be the same length")
-    )
-  }
-  if (anyNA(s)) {
-    stop(
-      gettext("'s' cannot contain NAs")
-    )
-  }
-  method <- match.arg(method)
-  ns <- tabulate(s, nbins = nlevels(s))
-  if (any(min > ns)) {
-    stop(
-      gettext("'min' must be smaller than the population size for each stratum")
-    )
-  }
-  N <- N - min * nlevels(s)
-  if (N < 0) {
-    stop(
-      gettext("minimal allocation is larger than 'N'")
-    )
-  }
-  # apportionment method
-  round <- if (method == "Largest-remainder") {
-    largest_remainder
-  } else {
-    highest_averages(method)
-  }
-  p <- vapply(split(x, s), sum, numeric(1L))
-  res <- structure(rep(min, length(p)), names = levels(s))
-  # redistribute sample units for those that cap out at the stratum size
-  repeat {
-    res <- round(p, N, res)
-    d <- pmax(res - ns, 0)
-    over <- as.logical(d)
-    if (!any(over)) break
-    res[over] <- ns[over]
-    p[over] <- 0  
-    N <- sum(d)
-  }
-  res
-}
 
 #---- Methods for class 'sps' ----
 # levels.default() extracts levels; no need for a method