+++ tests/test-repweights.R
@@ -3,9 +3,14 @@ library(sps)
 set.seed(1234)
 
 # Corner cases
-# unclass(sps_repweights(weights(sps(1:10, 0)), 10))
-unclass(sps_repweights(weights(sps(1:10, 10)), 5, tau = 3))
-unclass(sps_repweights(weights(sps(1:10, 10)), 10, dist = rnorm))
+all.equal(
+  unclass(sps_repweights(weights(sps(1:10, 10)), 5, tau = 3)),
+  structure(matrix(1, 10, 5), tau = 3)
+)
+all.equal(
+  unclass(sps_repweights(weights(sps(1:10, 10)), 10, dist = rnorm)),
+  structure(matrix(1, 10, 10), tau = 1)
+)
 
 # All rep weights should be positive with the TA strata having all 1s
 all(sps_repweights(1:5, tau = 2) > 0)